+++ events/filters.go
@@ -13,6 +13,22 @@ import (
 
 // checkEventFilters processes incoming event through configured filters to determine if it should generate a notification
 func checkEventFilters(event models.Event) bool {
+	// Drop event if no snapshot or clip is available - Event is likely being filtered on Frigate side.
+	// For example, if a camera has `required_zones` set - then there may not be any clip or snap until
+	// object moves into required zone
+	if !event.HasClip && !event.HasSnapshot {
+		log.Info().
+			Str("event_id", event.ID).
+			Msg("Event dropped - No snapshot or clip available")
+		return false
+	}
+	// Drop event if no snapshot & skip_nosnap is true
+	if !event.HasSnapshot && strings.ToLower(config.ConfigData.Alerts.General.NoSnap) == "drop" {
+		log.Info().
+			Str("event_id", event.ID).
+			Msg("Event dropped - No snapshot available")
+		return false
+	}
 	// Check quiet hours
 	if isQuietHours() {
 		log.Info().