+++ docs/changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [v0.3.5](https://github.com/0x2142/frigate-notify/releases/tag/v0.3.5) - Upcoming Release
+ - Fixed issue where built-in alert templates were not being included in binary releases
+ - Fixed issue where a notification may not be sent if previous event update from Frigate did not contain a snapshot
+ - Added `from` & `ignoressl` config items to `smtp` notifier
+ - Added ability to send webhook notifications via HTTP GET requests
+ - Added support for URL parameters with webhook notifications
+
 ## [v0.3.4](https://github.com/0x2142/frigate-notify/releases/tag/v0.3.4) - Aug 15 2024
  - Fixed issue where `unzoned: drop` config would drop all notifications